Preliminary Resource Allocation
Set up global vars for resources and upgrades to the ship, as well as the ability to allocate resources to parts of ship using sliders, and corresponding buttons for even distribution and clearing slider values.

Sliders can't increase if the total amount exceeds resources available, and even allocation accounts for division with remainders!