Added Encounter Map, Shop Screen, Main Menu, Credits, Settings Screen
Added basic functionality for Encounter Map: User can travel to a battle or the shop.
Added basic main menu functionality:
User can start the game, go to the settings screen, go to the credit screen, or quit the application.
Added bare bones shop functionality: User can access the shop screen and have a prompt one message from the shop keeper or can return to the encounter map.
Added settings screen with no current functionality other than returning the player to the main menu.
Added the credits screen with the most basic credits and the ability for the player to return to the main menu.